# docs/openapi.json
{
 "openapi": "3.0.1",
 "info": {
  "title": "OpenAPI definition",
  "version": "v0"
 },
 "servers": [
  {
   "url": "http://localhost:18090",
   "description": "Generated server url"
  }
 ],
 "paths": {
...
  "/api/v1/messages": {
   "post": {
    "tags": [
     "api-controller"
    ],
    "summary": "Demander l'envoi d'un MT (idempotent via clientRef)",
    "operationId": "send",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SendRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "*/*": {
        "schema": {
         "$ref": "#/components/schemas/MessageView"
        }
       }
      }
     }
    }
   }
  },
...
  "/admin/reconciliation/{operatorCode}": {
   "post": {
    "tags": [
     "admin-controller"
    ],
    "operationId": "reconcile",
    "parameters": [
     {
      "name": "operatorCode",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string"
      }
     },
     {
      "name": "separator",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "default": ";"
      }
     },
     {
      "name": "from",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "format": "date-time"
      }
     },
     {
      "name": "to",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "format": "date-time"
      }
     },
     {
      "name": "idColumn",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "default": "event_id"
      }
     },
     {
      "name": "amountColumn",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "default": "amount"
      }
     },
     {
      "name": "statusColumn",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "default": "status"
      }
     }
    ],
    "requestBody": {
     "content": {
      "multipart/form-data": {
       "schema": {
        "required": [
         "file"
        ],
        "type": "object",
        "properties": {
         "file": {
          "type": "string",
          "format": "binary"
         }
        }
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "*/*": {
        "schema": {
         "type": "object",
         "additionalProperties": {
          "type": "object"
         }
        }
       }
      }
     }
    }
   }
  },
...
  "/admin/reports/summary/export": {
   "get": {
    "tags": [
     "reporting-query-controller"
    ],
    "operationId": "summaryExport",
    "parameters": [
     {
      "name": "format",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "default": "pdf"
      }
     },
     {
      "name": "hours",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "format": "int32",
       "default": 24
      }
     }
    ],
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "*/*": {
        "schema": {
         "type": "string",
         "format": "byte"
        }
       }
      }
     }
    }
   }
  },
  "/admin/reconciliation/{batch}": {
   "get": {
    "tags": [
     "reporting-query-controller"
    ],
    "operationId": "reconItems",
    "parameters": [
     {
      "name": "batch",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string"
      }
     },
     {
      "name": "result",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "enum": [
        "MATCHED",
        "AMOUNT_MISMATCH",
        "MISSING_ON_PLATFORM",
        "MISSING_ON_OPERATOR",
        "STATUS_MISMATCH"
       ]
      }
     },
     {
      "name": "page",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "format": "int32"
      }
     },
     {
      "name": "size",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "format": "int32"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "*/*": {
        "schema": {
         "type": "array",
         "items": {
          "type": "object",
          "additionalProperties": {
           "type": "object"
          }
         }
        }
       }
      }
     }
    }
   }
  },
  "/admin/reconciliation/{batch}/export": {
   "get": {
    "tags": [
     "reporting-query-controller"
    ],
    "operationId": "reconExport",
    "parameters": [
     {
      "name": "batch",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string"
      }
     },
     {
      "name": "format",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "default": "csv"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "*/*": {
        "schema": {
         "type": "string",
         "format": "byte"
        }
       }
      }
     }
    }
   }
  },
...
  "/admin/messages": {
   "get": {
    "tags": [
     "reporting-query-controller"
    ],
    "operationId": "messages",
    "parameters": [
     {
      "name": "id",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string"
      }
     },
     {
      "name": "msisdn",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string"
      }
     },
     {
      "name": "operator",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string"
      }
     },
     {
      "name": "shortCode",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string"
      }
     },
     {
      "name": "status",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "enum": [
        "PENDING",
        "SUBMITTED",
        "DELIVERED",
        "EXPIRED",
        "UNDELIVERABLE",
        "REJECTED",
        "UNKNOWN",
        "FAILED"
       ]
      }
     },
     {
      "name": "from",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "format": "date-time"
      }
     },
     {
      "name": "to",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "format": "date-time"
      }
     },
     {
      "name": "page",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "format": "int32"
      }
     },
     {
      "name": "size",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "format": "int32"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "*/*": {
        "schema": {
         "type": "array",
         "items": {
          "type": "object",
          "additionalProperties": {
           "type": "object"
          }
         }
        }
       }
      }
     }
    }
   }
  },
...
  "/admin/ledger": {
   "get": {
    "tags": [
     "reporting-query-controller"
    ],
    "operationId": "ledger",
    "parameters": [
     {
      "name": "status",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "enum": [
        "PENDING",
        "ACCEPTED",
        "CHARGED",
        "REJECTED",
        "REVERSED",
        "DISPUTED"
       ]
      }
     },
     {
      "name": "from",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "format": "date-time"
      }
     },
     {
      "name": "to",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "format": "date-time"
      }
     },
     {
      "name": "page",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "format": "int32"
      }
     },
     {
      "name": "size",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "format": "int32"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "*/*": {
        "schema": {
         "type": "array",
         "items": {
          "type": "object",
          "additionalProperties": {
           "type": "object"
          }
         }
        }
       }
      }
     }
    }
   }
  },
  "/admin/ledger/export": {
   "get": {
    "tags": [
     "reporting-query-controller"
    ],
    "operationId": "ledgerExport",
    "parameters": [
     {
      "name": "format",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "default": "csv"
      }
     },
     {
      "name": "status",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "enum": [
        "PENDING",
        "ACCEPTED",
        "CHARGED",
        "REJECTED",
        "REVERSED",
        "DISPUTED"
       ]
      }
     },
     {
      "name": "from",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "format": "date-time"
      }
     },
     {
      "name": "to",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "format": "date-time"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "*/*": {
        "schema": {
         "type": "string",
         "format": "byte"
        }
       }
      }
     }
    }
   }
  },
  "/admin/dashboard": {
   "get": {
    "tags": [
     "reporting-query-controller"
    ],
    "operationId": "dashboard",
    "parameters": [
     {
      "name": "hours",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "format": "int32",
       "default": 24
      }
     }
    ],
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "*/*": {
        "schema": {
         "type": "object",
         "additionalProperties": {
          "type": "object"
         }
        }
       }
      }
     }
    }
   }
  },
  "/admin/consents": {
   "get": {
    "tags": [
     "reporting-query-controller"
    ],
    "operationId": "consents",
    "parameters": [
     {
      "name": "msisdn",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string"
      }
     },
     {
      "name": "serviceId",
      "in": "query",
      "required": true,
      "schema": {
       "type": "integer",
       "format": "int64"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "*/*": {
        "schema": {
         "type": "array",
         "items": {
          "type": "object",
          "additionalProperties": {
           "type": "object"
          }
         }
        }
       }
      }
     }
    }
   }
  },
  "/admin/consents/export": {
   "get": {
    "tags": [
     "reporting-query-controller"
    ],
    "operationId": "consentsExport",
    "parameters": [
     {
      "name": "msisdn",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string"
      }
     },
     {
      "name": "serviceId",
      "in": "query",
      "required": true,
      "schema": {
       "type": "integer",
       "format": "int64"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "text/csv": {
        "schema": {
         "type": "string",
         "format": "byte"
        }
       }
      }
     }
    }
   }
  },
...
  "/admin/audit": {
   "get": {
    "tags": [
     "reporting-query-controller"
    ],
    "operationId": "audit",
    "parameters": [
     {
      "name": "page",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "format": "int32"
      }
     },
     {
      "name": "size",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "format": "int32"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "*/*": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/AuditLog"
         }
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "SendRequest": {
    "required": [
     "text",
     "to"
    ],
    "type": "object",
    "properties": {
     "to": {
      "type": "string"
     },
     "text": {
      "type": "string"
     },
     "sender": {
      "type": "string"
     },
     "serviceId": {
      "type": "integer",
      "format": "int64"
     },
     "operator": {
      "type": "string"
     },
     "priority": {
      "type": "string",
      "enum": [
       "TRANSACTIONAL",
       "CONFIRMATION",
       "BULK"
      ]
     },
     "clientRef": {
      "type": "string"
     },
     "scheduleAt": {
      "type": "string",
      "format": "date-time"
     }
    }
   },
   "MessageView": {
    "type": "object",
    "properties": {
     "id": {
      "type": "string"
     },
     "clientRef": {
      "type": "string"
     },
     "to": {
      "type": "string"
     },
     "status": {
      "type": "string"
     },
     "rawStatus": {
      "type": "string"
     },
     "segments": {
      "type": "integer",
      "format": "int32"
     },
     "encoding": {
      "type": "string"
     },
     "history": {
      "type": "array",
      "items": {
       "type": "object",
       "additionalProperties": {
        "type": "object"
       }
      }
     }
    }
   },
...
   "AuditLog": {
    "type": "object",
    "properties": {
     "id": {
      "type": "integer",
      "format": "int64"
     },
     "actor": {
      "type": "string"
     },
     "action": {
      "type": "string"
     },
     "target": {
      "type": "string"
     },
     "detail": {
      "type": "string"
     },
     "at": {
      "type": "string",
      "format": "date-time"
     }
    }
   }
  }
 }
}